refactor: move all optics-specific behavior into folder
