+++ typescript/optics-provider/src/optics/index.ts
@@ -1,0 +1,4 @@
+export { BridgeContracts } from './BridgeContracts';
+export { CoreContracts } from './CoreContracts';
+export { OpticsDomain, mainnetDomains } from './OpticsDomain';
+export { OpticsContext, mainnet } from './OpticsContext';
